Implemented operator class is working as intended, fixed bug when revealing mines at game win, also added functionality for counting and displaying the remaining potential mines to be flagged.

+++ BuscaMinas.Models/Player.cs
@@ -6,6 +6,6 @@ namespace BuscaMinas.Models
     {
         public int Id { get; set; }
         public string? Name { get; set; }
-        public int HighScore { get; set; }
+        public decimal HighScore { get; set; }
     }
 }
